refactor: deduplicate GraphQL query and computePrWeight into @chapa/shared
- Move CONTRIBUTION_QUERY to packages/shared/src/github-query.ts
- Move computePrWeight() to packages/shared/src/scoring.ts
- Move RawContributionData type to packages/shared/src/types.ts
- Update web app and CLI to import from @chapa/shared
- Add 11 new tests for shared modules (387 total)
