+++ apps/web/lib/github/queries.ts
@@ -1,98 +1,8 @@
-// ---------------------------------------------------------------------------
-// Types for raw GitHub GraphQL response data
-// ---------------------------------------------------------------------------
+import type { RawContributionData } from "@chapa/shared";
+import { CONTRIBUTION_QUERY } from "@chapa/shared";
 
-export interface RawContributionData {
-  login: string;
-  name: string | null;
-  avatarUrl: string;
-  contributionCalendar: {
-    totalContributions: number;
-    weeks: {
-      contributionDays: {
-        date: string;
-        contributionCount: number;
-      }[];
-    }[];
-  };
-  pullRequests: {
-    totalCount: number;
-    nodes: {
-      additions: number;
-      deletions: number;
-      changedFiles: number;
-      merged: boolean;
-    }[];
-  };
-  reviews: { totalCount: number };
-  issues: { totalCount: number };
-  repositories: {
-    totalCount: number;
-    nodes: {
-      nameWithOwner: string;
-      defaultBranchRef: {
-        target: { history: { totalCount: number } };
-      } | null;
-    }[];
-  };
-}
-
-// ---------------------------------------------------------------------------
-// GraphQL query
-// ---------------------------------------------------------------------------
-
-const CONTRIBUTION_QUERY = `
-query($login: String!, $since: DateTime!, $until: DateTime!, $historySince: GitTimestamp!, $historyUntil: GitTimestamp!) {
-  user(login: $login) {
-    login
-    name
-    avatarUrl
-    contributionsCollection(from: $since, to: $until) {
-      contributionCalendar {
-        totalContributions
-        weeks {
-          contributionDays {
-            date
-            contributionCount
-          }
-        }
-      }
-      pullRequestContributions(first: 100) {
-        totalCount
-        nodes {
-          pullRequest {
-            additions
-            deletions
-            changedFiles
-            merged
-          }
-        }
-      }
-      pullRequestReviewContributions(first: 1) {
-        totalCount
-      }
-      issueContributions(first: 1) {
-        totalCount
-      }
-    }
-    repositories(first: 20, ownerAffiliations: [OWNER, COLLABORATOR], orderBy: {field: PUSHED_AT, direction: DESC}) {
-      totalCount
-      nodes {
-        nameWithOwner
-        defaultBranchRef {
-          target {
-            ... on Commit {
-              history(since: $historySince, until: $historyUntil) {
-                totalCount
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-`;
+// Re-export for consumers that import from this module
+export type { RawContributionData };
 
 // ---------------------------------------------------------------------------
 // Fetch function
