+++ apps/web/lib/github/stats90d.ts
@@ -1,22 +1,6 @@
 import type { Stats90d, HeatmapDay } from "@chapa/shared";
+import { computePrWeight } from "@chapa/shared";
 import { fetchContributionData } from "./queries";
-
-// ---------------------------------------------------------------------------
-// PR weight formula: w = 0.5 + 0.25*ln(1+filesChanged) + 0.25*ln(1+additions+deletions)
-// Capped at 3.0 per PR. Total capped at 40.
-// ---------------------------------------------------------------------------
-
-function computePrWeight(pr: {
-  additions: number;
-  deletions: number;
-  changedFiles: number;
-}): number {
-  const w =
-    0.5 +
-    0.25 * Math.log(1 + pr.changedFiles) +
-    0.25 * Math.log(1 + pr.additions + pr.deletions);
-  return Math.min(w, 3.0);
-}
 
 // ---------------------------------------------------------------------------
 // fetchStats90d — main aggregation function
