+++ packages/cli/src/fetch-emu.ts
@@ -1,78 +1,5 @@
 import type { Stats90d, HeatmapDay } from "@chapa/shared";
-
-// ---------------------------------------------------------------------------
-// GraphQL query (duplicated from apps/web/lib/github/queries.ts)
-// This is ~stable code; sharing via a common package is a future refactor.
-// ---------------------------------------------------------------------------
-
-const CONTRIBUTION_QUERY = `
-query($login: String!, $since: DateTime!, $until: DateTime!, $historySince: GitTimestamp!, $historyUntil: GitTimestamp!) {
-  user(login: $login) {
-    login
-    name
-    avatarUrl
-    contributionsCollection(from: $since, to: $until) {
-      contributionCalendar {
-        totalContributions
-        weeks {
-          contributionDays {
-            date
-            contributionCount
-          }
-        }
-      }
-      pullRequestContributions(first: 100) {
-        totalCount
-        nodes {
-          pullRequest {
-            additions
-            deletions
-            changedFiles
-            merged
-          }
-        }
-      }
-      pullRequestReviewContributions(first: 1) {
-        totalCount
-      }
-      issueContributions(first: 1) {
-        totalCount
-      }
-    }
-    repositories(first: 20, ownerAffiliations: [OWNER, COLLABORATOR], orderBy: {field: PUSHED_AT, direction: DESC}) {
-      totalCount
-      nodes {
-        nameWithOwner
-        defaultBranchRef {
-          target {
-            ... on Commit {
-              history(since: $historySince, until: $historyUntil) {
-                totalCount
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
-`;
-
-// ---------------------------------------------------------------------------
-// PR weight formula (duplicated from apps/web/lib/github/stats90d.ts)
-// ---------------------------------------------------------------------------
-
-function computePrWeight(pr: {
-  additions: number;
-  deletions: number;
-  changedFiles: number;
-}): number {
-  const w =
-    0.5 +
-    0.25 * Math.log(1 + pr.changedFiles) +
-    0.25 * Math.log(1 + pr.additions + pr.deletions);
-  return Math.min(w, 3.0);
-}
+import { CONTRIBUTION_QUERY, computePrWeight } from "@chapa/shared";
 
 // ---------------------------------------------------------------------------
 // Fetch EMU stats via GraphQL (requires EMU token with auth)
