+++ packages/shared/src/github-query.ts
@@ -1,0 +1,65 @@
+/**
+ * GitHub GraphQL query for fetching a user's contribution data over 90 days.
+ *
+ * Variables:
+ * - $login: String! — GitHub username
+ * - $since: DateTime! — Start of window (contributionsCollection)
+ * - $until: DateTime! — End of window (contributionsCollection)
+ * - $historySince: GitTimestamp! — Start of window (commit history)
+ * - $historyUntil: GitTimestamp! — End of window (commit history)
+ *
+ * Note: DateTime and GitTimestamp are different GraphQL types but accept
+ * the same ISO 8601 strings. They must be declared as separate variables.
+ */
+export const CONTRIBUTION_QUERY = `
+query($login: String!, $since: DateTime!, $until: DateTime!, $historySince: GitTimestamp!, $historyUntil: GitTimestamp!) {
+  user(login: $login) {
+    login
+    name
+    avatarUrl
+    contributionsCollection(from: $since, to: $until) {
+      contributionCalendar {
+        totalContributions
+        weeks {
+          contributionDays {
+            date
+            contributionCount
+          }
+        }
+      }
+      pullRequestContributions(first: 100) {
+        totalCount
+        nodes {
+          pullRequest {
+            additions
+            deletions
+            changedFiles
+            merged
+          }
+        }
+      }
+      pullRequestReviewContributions(first: 1) {
+        totalCount
+      }
+      issueContributions(first: 1) {
+        totalCount
+      }
+    }
+    repositories(first: 20, ownerAffiliations: [OWNER, COLLABORATOR], orderBy: {field: PUSHED_AT, direction: DESC}) {
+      totalCount
+      nodes {
+        nameWithOwner
+        defaultBranchRef {
+          target {
+            ... on Commit {
+              history(since: $historySince, until: $historyUntil) {
+                totalCount
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+`;
